# server/docs/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  epic.TrashEpic:
    properties:
      color:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      item_count:
        type: integer
      name:
        type: string
    type: object
...
  project.TrashProject:
    properties:
      color:
        type: string
      deleted_at:
        type: string
      epic_count:
        type: integer
      id:
        type: string
      item_count:
        type: integer
      name:
        type: string
      prefix:
        type: string
    type: object
...
  sprint.CloseResponse:
    properties:
      next_sprint:
        $ref: '#/definitions/sprint.Sprint'
      spilled_count:
        type: integer
    type: object
  sprint.CreateRequest:
    properties:
      start_date:
        description: StartDate is optional; defaults to today if omitted.
        type: string
    type: object
  sprint.ListResponse:
    properties:
      sprints:
        items:
          $ref: '#/definitions/sprint.Sprint'
        type: array
    type: object
  sprint.SpilloverPreviewResponse:
    properties:
      count:
        type: integer
    type: object
  sprint.Sprint:
    properties:
      board_id:
        type: string
      created_at:
        type: string
      days_remaining:
        type: integer
      days_total:
        type: integer
      end_date:
        type: string
      id:
        type: string
      name:
        type: string
      sprint_number:
        type: integer
      start_date:
        type: string
      state:
        type: string
    type: object
  trash.TrashResponse:
    properties:
      epics:
        items:
          $ref: '#/definitions/epic.TrashEpic'
        type: array
      projects:
        items:
          $ref: '#/definitions/project.TrashProject'
        type: array
    type: object
...
host: api.cumin.ashinsabu.com
info:
  contact: {}
  description: OSS sprint board API
  title: Cumin API
  version: "1.0"
paths:
...
  /api/sprints:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sprint.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List sprints
      tags:
      - sprints
    post:
      consumes:
      - application/json
      parameters:
      - description: Sprint start date (optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/sprint.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sprint.Sprint'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create sprint
      tags:
      - sprints
  /api/sprints/{id}/activate:
    post:
      parameters:
      - description: Sprint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sprint.Sprint'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Activate sprint
      tags:
      - sprints
  /api/sprints/{id}/close:
    post:
      parameters:
      - description: Sprint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sprint.CloseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Close sprint
      tags:
      - sprints
  /api/sprints/{id}/spillover-preview:
    get:
      parameters:
      - description: Sprint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sprint.SpilloverPreviewResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Spillover preview
      tags:
      - sprints
  /api/sprints/active:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sprint.Sprint'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get active sprint
      tags:
      - sprints
  /api/trash:
    delete:
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Empty trash
      tags:
      - trash
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trash.TrashResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List trash
      tags:
      - trash
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"